requirements.txt
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.1.2
numpy==1.26.4
pandas==2.2.0
plotly==5.21.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
Werkzeug==3.0.1